+++ src/Symfony/Component/HttpClient/Recorder/Store/FilesystemStore.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Symfony\Component\HttpClient\Recorder\Store;
 
 use Symfony\Component\Filesystem\Filesystem;
@@ -12,11 +21,10 @@ final class FilesystemStore implements StoreInterface
 {
     private Filesystem $filesystem;
 
-    public function __construct(?Filesystem $filesystem)
+    public function __construct(?Filesystem $filesystem = null)
     {
         if (null === $filesystem && !class_exists(Filesystem::class)) {
             throw new \LogicException(\sprintf('The "%s" handler needs a Filesystem. Try running "composer require symfony/filesystem".', __CLASS__));
-
         }
         $this->filesystem = $filesystem ?? new Filesystem();
     }