+++ src/Symfony/Component/HttpClient/Recorder/Store/StoreInterface.php
@@ -1,7 +1,15 @@
 <?php
 
-namespace Symfony\Component\HttpClient\Recorder\Store;
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
+namespace Symfony\Component\HttpClient\Recorder\Store;
 
 use Symfony\Component\HttpClient\Har\HarFile;